moved circuitbreaker to lib folder
